+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Added `piral-react` as converter for React
 - Added possibility to publish emulator as a website (#644)
 - Added orchestration engine choices (`systemjs`, `module-federation`) (#643)
+- Added isolation mode option to opt-in for `piral-component` boundary
 - Added option to specify runtime execution path for bundler plugins
 
 ## 1.3.3 (November 2, 2023)